Move processing of animation loop into game engine
Time and renderer are passed in, so should be easy to test (except
it calls the renderer, which will mean mocking, Hmmm.

We should keep animation loops and game tick events triggering from
outside of the game engine, to keep it deterministic and stateless.

Also I guess external event queues (or something) should be external.