Random changes as found April 2023
Appears to include changes to how modules are loaded, but I'm not sure.
I'm on an aeroplane with no internet so can't check.
In my head - I want to ditch the electron and turn this into a browser
application, I'm struggling to do this on the plane as I'm rusty
and can't find a functioning local http server that's installed.